Split out Composer to a separately named type
